Refactor constants fields
