+++ ThirdPartyAdapters/liftoffmonetize/.gitignore
@@ -1,0 +1,7 @@
+*.iml
+.gradle
+/local.properties
+/.idea
+.DS_Store
+*/build
+/captures
