+++ ThirdPartyAdapters/liftoffmonetize/liftoffmonetize/src/main/java/com/google/ads/mediation/vungle/VungleConstants.java
@@ -1,0 +1,10 @@
+package com.google.ads.mediation.vungle;
+
+public class VungleConstants {
+
+  public static final String KEY_APP_ID = "appid";
+  public static final String KEY_USER_ID = "userId";
+  public static final String KEY_ORIENTATION = "adOrientation";
+  public static final String KEY_PLAY_PLACEMENT = "playPlacement";
+
+}
